Move data stores to separately queueable scripts
