+++ inc/blocks/personalization/index.js
@@ -2,9 +2,6 @@ import blockData from './block.json';
 import edit from './edit';
 import save from './save';
 
-// Load analytics data store.
-import './data/analytics';
-
 const { registerBlockType } = wp.blocks;
 const { __ } = wp.i18n;
