+++ src/audiences/ui/components/group.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
 
-import { defaultRule } from '../data/defaults';
+import { defaultRule } from '../../data/defaults';
 
 import Rule from './rule';
 import SelectInclude from './select-include';
